backend/pyproject.toml
[tool.poetry]
name = "chillmap-backend"
version = "0.1.0"
description = ""
authors = ["Denis Kutov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
django = "^4.2.1"
psycopg2-binary = "^2.9.6"
django-cors-headers = "^3.14.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
autoflake = "^2.1.1"
django-debug-toolbar = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
